+++ src/effector-persist.ts
@@ -2,18 +2,18 @@ import { createEvent } from "effector";
 
 type IPersistConfig = {
   key?: string;
-}
+};
 
 const defaultConfig = {
-  key: 'persist'
+  key: "persist"
 };
 
 export const withPersist = <State>(store: any, config: IPersistConfig = defaultConfig) => {
   const name = store.shortName;
   const persistKey = `${config.key}:${name}`;
-  const rehydrate = createEvent('@PERSIST/REHYDRATE');
+  const rehydrate = createEvent("@PERSIST/REHYDRATE");
 
-  chrome.storage.sync.get([persistKey], (result) => {
+  chrome.storage.sync.get([persistKey], result => {
     if (result[persistKey]) {
       store.on(rehydrate, () => JSON.parse(result[persistKey]));
       rehydrate();
@@ -21,9 +21,7 @@ export const withPersist = <State>(store: any, config: IPersistConfig = defaultC
   });
 
   store.watch((state: any) => {
-    chrome.storage.sync.set({ [persistKey]: JSON.stringify(state) }, () => {
-      console.log('Value is set to ', JSON.stringify(state));
-    });
+    chrome.storage.sync.set({ [persistKey]: JSON.stringify(state) });
   });
 
   return store;